CHROMIUM: iwl7000: pcie: remove allocator pool
9000 hardware device supports multiple rx queues, and RBDs
may be moved by the hardware to serve another queues. A RBD
is assigned with an internal index (vid) so when a RB is closed
we can know which RBD it is.
However, my previous patch neglected to assign the vid to
the allocator's initial pool, with the devastating result of
treating the RB with vid 0 as closed multiple times while
it is actually used by the hardware.
Given how easily it was overlooked, remove the allocator pool
altogether. Instead, init the allocator empty_rbd list with
the minimum required RBDs from the default queue's pool. Move
the vid initialization to the same location.

Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: ab79dbb7a2cbeb9d8cf7a6af571ef2afdbbf4703