UPSTREAM: usb: dwc3: Update maximum_speed for SuperSpeedPlus
If the maximum_speed is not set, set it to a known value, either
SuperSpeed or SuperSpeedPlus based on the type of controller we are
using. If we are on DWC_usb31 controller, check the PHY interface to see
if it is capable of SuperSpeedPlus.

Also this check is moved after dwc3_core_init() so that we can check
dwc->revision.

Signed-off-by: John Youn <johnyoun@synopsys.com>
Signed-off-by: Felipe Balbi <balbi@kernel.org>
(cherry picked from commit 2c7f1bd9127a1a49ee25d9c2b2ce17b11c7fb05f)

BUG=chrome-os-partner:52744
TEST=Build and boot

Change-Id: I150210199f1a80e47a71a86cf2217bf4b740436f
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/340621
Reviewed-by: Douglas Anderson <dianders@chromium.org>