ANDROID: vfs: add d_canonical_path for stacked filesystem support
Inotify does not currently know when a filesystem
is acting as a wrapper around another fs. This means
that inotify watchers will miss any modifications to
the base file, as well as any made in a separate
stacked fs that points to the same file.
d_canonical_path solves this problem by allowing the fs
to map a dentry to a path in the lower fs. Inotify
can use it to find the appropriate place to watch to
be informed of all changes to a file.

Signed-off-by: Daniel Rosenberg <drosen@google.com>
(cherry picked from https://android.googlesource.com/kernel/common android-4.4
 commit 530d9c5fc268b0c90ffbca40384f9ee657076724)

BUG=chromium:612862
TEST=build and boot on a few devices

Change-Id: I09563baffad1711a045e45c1bd0bd8713c2cc0b6
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/345856
Reviewed-by: Guenter Roeck <groeck@chromium.org>