UPSTREAM: drm/i915: Move the encoder vs. FDI dotclock check out from encoder .get_config()
Currently we check if the encoder's idea of dotclock agrees with what
we calculated based on the FDI parameters. We do this in the encoder
.get_config() hooks, which isn't so nice in case the BIOS (or some other
outside party) made a mess of the state and we're just trying to take
over.

So as a prep step to being able sanitize such a bogus state, move the
the sanity check to just after we've read out the entire state. If
we then need to sanitize a bad state, it should be easier to move the
sanity check to occur after sanitation instead of before it.

Signed-off-by: Ville Syrjl <ville.syrjala@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1455738073-14502-3-git-send-email-ville.syrjala@linux.intel.com
Reviewed-by: Imre Deak <imre.deak@intel.com>
(cherry picked from commit e3b247da89f0adb04010f9dad6a3e41ed085d24e
from drm-intel-next-2016-03-14)
Signed-off-by: Tarun Vyas <tarun.vyas@intel.com>

BUG=chrome-os-partner:52640
TEST=Image builds, boots and is browse-able

Change-Id: Ic6b595dbbee7af692c67a0cd46df7db029eab2b6
Reviewed-on: https://chromium-review.googlesource.com/344970
Commit-Ready: Tarun Vyas <tarun.vyas@intel.com>
Tested-by: Russ Sage <russ.sage@intel.com>
Reviewed-by: Benson Leung <bleung@google.com>