CHROMIUM: iwl7000: mac80211: Fix BW upgrade for TDLS peers
It is possible that the station is connected to an AP
with bandwidth of 80+80MHz or 160MHz. In such cases
there is no need to perform an upgrade as the maximal
supported bandwidth is 80MHz.

In addition, when upgrading and setting center_freq1
and bandwidth to 80MHz also set center_freq2 to 0.

Signed-off-by: Ilan Peer <ilan.peer@intel.com>
iwl7000-tree: 35cb2c556748b61be27e0e620d525ca8fa6b58e6