UPSTREAM: ALSA: hda - Fix spurious kernel WARNING on Baytrail HDMI
snd_hdac_sync_audio_rate() call is mandatory only for HSW and later
models, but we call the function unconditionally blindly assuming that
the function doesn't do anything harmful.  But since recently, the
function checks the validity of the passed pin NID, and eventually
spews the warning if an unexpected pin is passed.  This is seen on old
chips like Baytrail.

The fix is to limit the call of this function again only for the chips
with the proper binding.  This can be identified by the same flag as
the eld notifier.

Reported-by: Ville Syrjl <ville.syrjala@linux.intel.com>
Tested-by: Ville Syrjl <ville.syrjala@linux.intel.com>
Cc: <stable@vger.kernel.org> # v4.5
Signed-off-by: Takashi Iwai <tiwai@suse.de>
From v4.6-rc2
(cherry picked from commit 93a9ff151754fbdf951b1b993bcf96453f6e36b3)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=chrome-os-partner:53967
TEST=with the whole series applied, Audio works on Amenia.

Change-Id: I43ce6558bda8ab9ac7d104a3166b99e597a49f5a
Reviewed-on: https://chromium-review.googlesource.com/347844
Commit-Ready: Vedang Patel <vedang.patel@intel.com>
Tested-by: Russ Sage <russ.sage@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>