UPSTREAM: backlight: lm3630a: explicitly apply PWM config extracted from pwm_args
Call pwm_apply_args() just after requesting the PWM device so that the
polarity and period are initialized according to the information
provided in pwm_args.

This is an intermediate state, and pwm_apply_args() should be dropped as
soon as the atomic PWM infrastructure is in place and the driver makes
use of it.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>
Signed-off-by: Thierry Reding <thierry.reding@gmail.com>
(cherry picked from git://git.kernel.org/pub/scm/linux/kernel/git/thierry.reding/linux-pwm.git for-next
 commit e466ff3ae121669e278ec20ed8ae55eaff0b4e5c)

BUG=chrome-os-partner:52666
TEST=build and boot for gru/kevin

Change-Id: I1eee4e08f4cc0480d208fb1ce1f8e0db8950cffe
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/345607
Reviewed-by: Stephen Barber <smbarber@chromium.org>