UPSTREAM: input: misc: max8997: Explicitly apply PWM config extracted from pwm_args
Call pwm_apply_args() just after requesting the PWM device so that the
polarity and period are initialized according to the information
provided in pwm_args.

This is an intermediate state, and pwm_apply_args() should be dropped as
soon as the atomic PWM infrastructure is in place and the driver makes
use of it.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>
Signed-off-by: Thierry Reding <thierry.reding@gmail.com>
(cherry picked from git://git.kernel.org/pub/scm/linux/kernel/git/thierry.reding/linux-pwm.git for-next
 commit e3b80f7ac66db189983ce5caa240be535b4613dc)

BUG=chrome-os-partner:52666
TEST=build and boot for gru/kevin

Change-Id: I21ca285c573a328b377c3a87972181f1f688af87
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/345608
Reviewed-by: Stephen Barber <smbarber@chromium.org>