UPSTREAM: drm: i915: Explicitly apply PWM config extracted from pwm_args
Call pwm_apply_args() just after requesting the PWM device so that the
polarity and period are initialized according to the information
provided in pwm_args.

This is an intermediate state, and pwm_apply_args() should be dropped as
soon as the atomic PWM infrastructure is in place and the driver makes
use of it.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>
Signed-off-by: Thierry Reding <thierry.reding@gmail.com>
(cherry picked from git://git.kernel.org/pub/scm/linux/kernel/git/thierry.reding/linux-pwm.git for-next
 commit f579b32b5960fe14473d57f0043474ff2b3ebe8a)

BUG=chrome-os-partner:52666
TEST=build and boot for gru/kevin

Change-Id: Ic63d2e42e5620ac1f36704fd0ff7a26e4b1d69e0
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/345610
Reviewed-by: Stephen Barber <smbarber@chromium.org>