UPSTREAM: ARM: Explicitly apply PWM config extracted from pwm_args
Call pwm_apply_args() just after requesting the PWM device so that the
polarity and period are initialized according to the information
provided in pwm_args.

This is an intermediate state, and pwm_apply_args() should be dropped as
soon as the atomic PWM infrastructure is in place and the driver makes
use of it.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>
Acked-by: Krzysztof Kozlowski <k.kozlowski@samsung.com>
Signed-off-by: Thierry Reding <thierry.reding@gmail.com>
(cherry picked from git://git.kernel.org/pub/scm/linux/kernel/git/thierry.reding/linux-pwm.git for-next
 commit b3d38e6cb693e737272e2971dc8c0ad400e25582)

BUG=chrome-os-partner:52666
TEST=build and boot for gru/kevin

Change-Id: Ibf60c56ccf761db17698934d3f457b4bca1df1f3
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/345611
Reviewed-by: Stephen Barber <smbarber@chromium.org>