ANDROID: fuse: Add support for d_canonical_path
Allows FUSE to report to inotify that it is acting
as a layered filesystem. The userspace component
returns a string representing the location of the
underlying file. If the string cannot be resolved
into a path, the top level path is returned instead.

bug: 23904372
Signed-off-by: Daniel Rosenberg <drosen@google.com>
(cherry picked from https://android.googlesource.com/kernel/common android-4.4
 commit 00ee836d89c5756b4e937eae58c9d0d1d7ffe560)

BUG=chromium:612862
TEST=build and boot on a few devices

Change-Id: Iabdca0bbedfbff59e9c820c58636a68ef9683d9f
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/345891
Reviewed-by: Guenter Roeck <groeck@chromium.org>