CHROMIUM: arm64: dts: rockchip: Enable UHS modes on kevin (not rev-0).
With all the latest timing patches, reworks, and the like, I've found
that my kevin-rev1 seems to work with UHS now.  Enable it.  We'll use a
drive strength for all lines of 8mA since that was decided over email to
be the best: not too much overshoot but cleaner signals.

Note: if you've got an old non-reworked kevin this might break you.
Presumably people on a rev-0 are likely not to have reworks but
hopefully everyone's rev-1 has reworks.  If you're having SD problems
after this change, have your board checked.

BUG=chrome-os-partner:53253
TEST=Use sdmmc on kevin.

Change-Id: I9a8cf5cd8d8694b2d18e3bd43849808fffec134f
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/343936
Reviewed-by: Brian Norris <briannorris@chromium.org>